requirements.txt
aiogram~=3.4
abacusai~=1.4
python-dotenv
backoff
